Injectable random number generator.
SecureRandom.getInstanceStrong().nextBytes() (used to generate an
initialisation vector) hangs on circleci. The build environment runs
out of entropy. This caused the unit tests to hang and ultimately
circleci fails the job after 10 minutes of no output. So to remedy
this allow a psuedo random number generator to be used when running
unit tests.