Ensure earlier processing errors don't crash the 'AfterStep' method.
An earlier error in the processing pipeline may mean that the final
buffer is not in a fit state to be written to s3. If this is the case the
final write may throw an exception which could mask the earlier error.

So, only perform the 'AfterStep' if the step completed successfully
and swallow (but log) any errors performed during the afterstep.