TRIGGER ADG (#177)
Trigger ADG on successful completion of export.

As each instance of HTME finishes it checks the status table to
see if the entire export has finished. If it has finished it sends
the monitoring message, furthermore if HTME determines that the
export has completed successfully it sends a message to the SNS
queue which ultimately results in ADG being triggered.

Tidied up the integration test set-up - removed unnecessary
containers, deduplicated the multitide of properties files.

Eliminated as many snyk violations as possible with a round of
dependency upgrades.