Add metrication.
Using the prometheus client library to add metrication counter,
guages and summaries.

Spring batch is metrics-aware and so its metrics have been enabled,
this gives counts such as records read, records processed, records
written, plus timings for the job as a whole and each of its steps.

Counters have been added manually for each of the alerts that we
currently trigger off of the cloudwatch logs.

Brought the README up-to-date.

Removed redundant spring profiles and redundant spring config from
the test classes (and removed their documentation).